일부 이미지 변경

+++ Slayer/Assets/00_Scripts/Skill/Skill04.cs
@@ -24,6 +24,6 @@ public class Skill04 : SkillBase
         //    maxHp
         //);
 
-        DamageText.Create(player.transform.position, amount, Color.green);
+        DamageText.Create(player.transform.position, (int)amount, Color.green);
     }
 }
